# dicom-manager/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/annotation/submissions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotation"
                ],
                "summary": "Получить сохранённую разметку",
                "parameters": [
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор задачи анализа",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "done",
                            "uncertain",
                            "skipped"
                        ],
                        "type": "string",
                        "description": "Статус разметки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Только актуальные отправки",
                        "name": "only_latest",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/dicom/upload": {
            "post": {
                "description": "Загрузить DICOM файл на сервер. Источник определяется ролью: admin — manual, specialist — clinic.",
                "consumes": [
                    "application/dicom"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dicom"
                ],
                "summary": "Загрузить DICOM файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Исходное имя файла: attachment; filename*=UTF-8''scan.dcm",
                        "name": "Content-Disposition",
                        "in": "header"
                    },
                    {
                        "description": "DICOM файл для загрузки.",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Файл успешно загружен.",
                        "schema": {
                            "$ref": "#/definitions/upload.UploadOut"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Нет доступа к файлу.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Обработка файла ещё не завершена.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dicom/upload/batch": {
            "post": {
                "description": "Загрузить несколько DICOM файлов. Файлы с активными задачами пропускаются; ответ содержит только новые задачи.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dicom"
                ],
                "summary": "Загрузить DICOM файлы батчами",
                "parameters": [
                    {
                        "type": "file",
                        "description": ".zip архив с DICOM файлами для загрузки.",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Файлы успешно загружены.",
                        "schema": {
                            "$ref": "#/definitions/upload_batch.UploadBatchOut"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Нет доступа к файлу.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "У всех файлов батча уже есть активные задачи.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Annotator": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "cols": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "rows": {
                    "type": "integer"
                }
            }
        },
...
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Record"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.Record": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "object"
                },
                "annotator": {
                    "$ref": "#/definitions/model.Annotator"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "image": {
                    "$ref": "#/definitions/model.Image"
                },
                "image_flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "job_id": {
                    "type": "string"
                },
                "schema_version": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "submission_id": {
                    "type": "string"
                },
                "superseded_by": {
                    "type": "string"
                },
                "supersedes": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "upload.UploadOut": {
            "type": "object",
            "properties": {
                "dicom_id": {
                    "type": "string"
                },
                "job_id": {
                    "description": "ID задачи на сервере для обработки DICOM файла.",
                    "type": "string"
                }
            }
        },
        "upload_batch.DicomData": {
            "type": "object",
            "properties": {
                "dicom_id": {
                    "type": "string"
                },
                "job_id": {
                    "description": "ID задачи на сервере для обработки DICOM файла.",
                    "type": "string"
                }
            }
        },
        "upload_batch.UploadBatchOut": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Список данных о DICOM файлах и задачах на сервере.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/upload_batch.DicomData"
                    }
                }
            }
        },
...
    }
}